pyproject.toml
[tool.poetry]
name = "speckle-blender"
version = "2.0.0"
description = "the Speckle 2.0 connector for Blender!"
authors = ["izzy lyseggen <izzy.lyseggen@gmail.com>", "Gergő Jedlicska <gergo@jedlicska.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = ">=3.8, <4.0.0"
specklepy = "^2.20.0"
attrs = "^23.1.0"

# [tool.poetry.group.local_specklepy.dependencies]
# specklepy = {path = "../specklepy", develop = true}

[tool.poetry.group.dev.dependencies]
fake-bpy-module-latest = "^20240524"
black = "23.11.0"
pylint = "^2.15.7"
ruff = "^0.4.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
